Change remove() so it returns false if the file does not exist.
Currently there is no way remove() can return false.  If the unlink fails it will throw
an exception.  Its better to give the user useful information about whether anything
was removed.

For #17

+++ lib/Path/Tiny.pm
@@ -680,18 +680,22 @@ sub relative { path( File::Spec->abs2rel( $_[0]->[PATH], $_[1] ) ) }
 For directories, this is like like calling C<remove_tree> from L<File::Path>.  An
 optional hash reference is passed through to C<remove_tree>.
 
-For files, the file is unlinked if it exists.  Unlike C<unlink>, if the file
-does not exist, this silently does nothing and returns a true value anyway.
+For files, the file is unlinked if it exists.
+
+If the path does not exist, it returns false.
 
 =cut
 
 sub remove {
     my ( $self, $opts ) = @_;
-    if ( -d $self->[PATH] ) {
+    if ( !-e $self->[PATH] ) {
+        return 0;
+    }
+    elsif ( -d $self->[PATH] ) {
         return File::Path::remove_tree( $self->[PATH], ref($opts) eq 'HASH' ? $opts : () );
     }
     else {
-        return ( -e $self->[PATH] ) ? unlink $self->[PATH] : 1;
+        return unlink $self->[PATH];
     }
 }
